+++ source/fitz/xml.c
@@ -991,9 +991,9 @@ static char *convert_to_utf8(fz_context *ctx, unsigned char *s, size_t n, int *d
 	{
 		fz_init_text_decoder(ctx, &dec, enc);
 		// NOTE: use decode_size if memory is more important than speed
-		m = dec.decode_bound(&dec, s, n);
+		m = (int)dec.decode_bound(&dec, s, (int)n);
 		dst = Memento_label(fz_malloc(ctx, m), "utf8");
-		dec.decode(&dec, dst, s, n);
+		dec.decode(&dec, dst, s, (int)n);
 		*dofree = 1;
 		return dst;
 	}
